+++ CHANGELOG.md
@@ -5,6 +5,7 @@ See [here](http://keepachangelog.com/) for the change log format.
 ## [in progress]
 
 - ADD: antecedent arguments to `implies` can be condensed type descriptions, not just predicates.
+- `all-built-like` and `<>all-built-like` add indexes to their error messages.
 
 ## [1.0-beta*]